Added protocol separation
The initial chaining key (kind of a hash) of protocols used to be all
zero.  This could make it difficult for parties from distinguishing one
protocol from another (for instance, the X and IK patterns start the
same way, and a server receiving an X message might initially accept it
even when it's configured for XK1).

By differentiating all protocols, we can now fail faster, perhaps even
engage in some protocol negotiation. (Proper protocol negotiation may
require authenticating a prelude, so I'm not sure this adds much.)

Users hell bent on protocol negotiation may XOR that initial chaining
key with the list of accepted protocol (or a hash thereof if it exceeds
32 bytes). I'm not sure I want to explicitly support that.

In any case, the change has practically zero overhead: instead of
initialising by zero, we initialise by a constant.