replaced "local/remote" by "client/server"
Using "local" and "remote" was convenient from the implementation point
of view, but it meant that the meaning of the arguments switched
depending on which function they were used with.

Very confusing.

With "client" and "server", the arguments always mean the same
thing. One less quirk for the user to learn.