Changed key derivation in the specs
The previous key derivation scheme didn't properly separate
protocols. It sure changed the derived keys, but it didn't guaranteed
that keys derived from different protocols with the same shared secrets
were completely unrelated: some chaining keys would differ only by a
known constant.

I personally conjecture that this was good enough: the Chacha20
permutation is such that flipping a bit of the key is essentially the
same as flipping a bit of the counter or nonce.  The conjecture is not
standard however, so I thought I'd best avoid it.

This patch changes two things: first, the protocol name is embedded in
the chaining hash. Second, chaining keys are taken from the *output* of
each chaining hash, instead of their input.  Here's the result(H0 is
HChacha20 with zero as the counter and nonce; Hp is HChacha20 with the
protocol name as counter and nonce):

   kex1            kex2            kex3
    |               |               |
  +----+          +----+          +----+
  | H0 |          | H0 |          | H0 |
  +----+          +----+          +----+
    |   +----+      |   +----+      |   +----+
    +---| Hp |--+--XOR--| Hp |--+--XOR--| Hp |--+
        +----+  |       +----+  |       +----+  |
               CK1             CK2             CK3

This makes sure that all chaining keys are independent from each other,
including in other protocol runs, even if users reuse ephemeral keys.
Good users would never commit such a deadly sin, but some clever fool
might be tempted to save a couple key exchanges during protocol
negotiation (which isn't supported, but y'know, users...).

Another possible benefit is that it allows us to extend Monokex to
support short preludes (under 8 bytes). This should be enough to do
protocol negotiation, where the number of protocols to chose from is
very limited (say, 8 or less).  Not that I actually intend to support
this (few users actually need it, and I'd rather avoid the complexity),
but someone desperate could XOR their prelude with bytes 0-7 of the
protocol name (no more, or we would risk internal nonce reuses), and use
the result as input for Hp.

TODO: update code generation.