Updated spec generation so they use XCKDF
XCKDF is the newest iteration of my Chacha20 based key derivation.  The
previous one was not bad, but it was delicate, and modifying it easily
lead to catastrophic nonce reuse.

This one uses a classic extract-then-expand scheme, and as such should
be more robust.  Most notably, the constant "pid" can now be any value,
including zero or one.