Check for end of input when parsing protocols
We previously stopped as soon as we didn't see a proper protocol.  This
silently stopped the parsing when a protocol name was missing.  The rest
of the code then stopped short, and only generated the first patterns,
or none if we omitted the first name.

Now we do check that the list of protocols ends with EOF, and the error
is no longer silent.