Added order check for pre-shared keys
Noise pattern don't naturally enforce the order of the pre-shared keys.
The KK pattern for instance starts with the following:

  -> s
  <- s
  ...

Strictly speaking, the following could be just as valid:

  <- s
  -> s
  ...

Noise (and Monokex) however just put the initiator's key first,
regardless of how it is specified in the pattern.  It is therefore not
clear that the order of the pattern has no influence on the end result.
It wouldn't influence security in any way, but it *would* change the
transcript hash.

To avoid this confusion, Monokex now enforces the order of the
pre-shared keys: if both are present, the initiator's key must be sent
first.  The error messages frames this as a limitation of Monokex (which
it is, but this limitation does not remove any functionality).