Corrected spec generation (wrong fallback hash)
The generated specs are modified in the following fashion:

- - __H9__  = Blake2b(H8  || Chacha20(K1, p2)) if msg2 has a payload, H8 otherwise
+ - __H9__  = Blake2b(H8  || Chacha20(K1, p2)) if msg2 has a payload, H7 otherwise

The error was, when an encrypted payload was absent, we referred to the
wrong hash. Instead of skipping one hash, we should have skipped two.
An encrypted payload eats up *two* hashes: one to generate the
encryption key, one to mix the encrypted payload into the transcript
hash.  Skipping only one hash forces us to generate an encryption key we
don't need.

Besides, the reference implementation already skips the two hashes.