Rewrite protocol specifications
The last version mixed the actions of the initiator and responder in a
way that may not be clear at all to an implementer.  This version is
much more explicit about which steps are taken, in which order.  Though
Monokex is based on a state machine, the specifications here describe
what it does directly, so users can just use the formulas instead of
coding the state machine.

Note that unlike tags, hashes and keys don't really have to be numbered:
each party can maintain a single hash and a single key, and overwrite it
at each update.  I numbered them to confirm I would be able to express
this in single assignment form, suitable for Verifpal models.